frontend_desktop/requirements.txt
customtkinter==5.2.2
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
